Working on theme picker
