+++ src/app/shared/components/theme-picker/theme-picker.component.scss
@@ -1,0 +1,19 @@
+.color-preview {
+  border-radius: 50%;
+  border: 1px solid;
+  width: 18px;
+  height: 18px;
+  margin-right: 15px;
+}
+
+.theme-name {
+  line-height: 1px;
+}
+
+.button-theme:hover {
+  color: var(--primary-color);
+}
+mat-icon {
+  width: 20px;
+  height: 20px;
+}
