+++ src/index.html
@@ -16,7 +16,7 @@
       rel="stylesheet"
     />
   </head>
-  <body class="mat-typography mat-app-background one-dark-theme">
+  <body class="mat-typography mat-app-background custom-theme">
     <app-root></app-root>
   </body>
 </html>
